Added -h and -v flags for plugin binaries, version is with git hash
