+++ plugins/aufs_cvmfs/VERSION
@@ -1,0 +1,1 @@
+0.1
